refactor(bun-dev): collapse resolve.ts into scaffold.ts
writeScaffold now takes cwd directly and resolves project files
internally. dev.ts and build.ts shed their resolveProject imports.
