+++ packages/bun-dev/src/dev.ts
@@ -1,13 +1,11 @@
-import { resolveProject } from "./resolve";
 import { writeScaffold } from "./scaffold";
 
 export async function runDev(cwd = process.cwd()): Promise<void> {
-  const resolved = await resolveProject(cwd);
-  const { servePath, bunfigPath } = writeScaffold(resolved);
+  const { servePath, bunfigPath } = writeScaffold(cwd);
 
   const child = Bun.spawn({
     cmd: ["bun", `--config=${bunfigPath}`, "--hot", servePath],
-    cwd: resolved.cwd,
+    cwd,
     env: process.env,
     stdio: ["inherit", "inherit", "inherit"],
   });
